Restructured project to run as command: uv run gh-profiler username
