+++ src/gh_profiler/__init__.py
@@ -1,0 +1,1 @@
+"""Make gh_profiler a package."""
