+++ src/gh_profiler/utils/analysis_utils.py
@@ -3,7 +3,7 @@
 from datetime import datetime as dt
 from datetime import timezone as tz
 
-from profile_data import profile_data as pdata
+from .profile_data import profile_data as pdata
 from . import flags
